Added Rough report and Term paper.

+++ src/ab/demo/other/ActionRobot.java
@@ -249,7 +249,7 @@ public class ActionRobot {
 			@Override
 			public int compare(Rectangle o1, Rectangle o2) {
 
-				return ((Integer)(o1.y)).compareTo((Integer)(o2.y));
+				return ((Integer)(o1.y)).compareTo(o2.y);
 
 			}	
 		});
