+++ src/ab/vision/VisionUtils.java
@@ -88,9 +88,9 @@ public class VisionUtils {
                 final int colourA = imgA.getRGB(x, y);
                 final int colourB = imgB.getRGB(x, y);
 
-				diff += Math.abs((int)((colourA & 0x00ff0000) >> 16) - (int)((colourB & 0x00ff0000) >> 16));
-				diff += Math.abs((int)((colourA & 0x0000ff00) >> 8) - (int)((colourB & 0x0000ff00) >> 8));
-				diff += Math.abs((int)(colourA & 0x000000ff) - (int)(colourB & 0x000000ff));
+				diff += Math.abs(((colourA & 0x00ff0000) >> 16) - ((colourB & 0x00ff0000) >> 16));
+				diff += Math.abs(((colourA & 0x0000ff00) >> 8) - ((colourB & 0x0000ff00) >> 8));
+				diff += Math.abs((colourA & 0x000000ff) - (colourB & 0x000000ff));
 
 
 			}
@@ -142,9 +142,9 @@ public class VisionUtils {
 		for (int y = 0; y < img.getHeight(); y++) {
 			for (int x = 0; x < img.getWidth(); x++) {
 				final int colour = img.getRGB(x, y);
-				sums[x] += (int) ((colour & 0x00ff0000) >> 16);
-				sums[x] += (int) ((colour & 0x0000ff00) >> 8);
-				sums[x] += (int) (colour & 0x000000ff);
+				sums[x] += (colour & 0x00ff0000) >> 16;
+				sums[x] += (colour & 0x0000ff00) >> 8;
+				sums[x] += colour & 0x000000ff;
 
 			}
 		}
